feat: products

+++ src/main/java/com/github/k1melo/crudapi/controller/ClientController.java
@@ -7,11 +7,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-import org.springframework.web.server.ResponseStatusException;
 
 import java.util.List;
-import java.util.Optional;
-import java.util.UUID;
 
 @RestController
 @RequestMapping(path = "/clients")
@@ -25,6 +22,11 @@ public class ClientController {
         return clientRepository.findAll();
     }
 
+    @GetMapping(path = "/{username}")
+    public ResponseEntity<String> getClientById(@PathVariable String username) {
+        if (clientRepository.findByUsername(username) == null) return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("User not found");
+        return ResponseEntity.status(200).body(clientRepository.findByUsername(username).getAll());
+    }
     @PostMapping
     public ResponseEntity<String> postClients(@RequestBody Client client) {
 
@@ -35,25 +37,27 @@ public class ClientController {
                 return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("Email not valid");
             }
             clientRepository.save(client);
-            return ResponseEntity.status(HttpStatus.CREATED).body("User registered: " + client.getName());
+            return ResponseEntity.status(HttpStatus.CREATED).body("User registered: " + client.getUsername());
         }
     }
 
-    @PutMapping(path = "/{id}")
-    public ResponseEntity<String> updateClients(@PathVariable Long id, @RequestBody Client client) {
+    @PutMapping(path = "/{username}")
+    public ResponseEntity<String> updateClientsById(@PathVariable String username, @RequestBody Client client) {
 
-        Client currentClient = this.clientRepository.findById(id).orElseThrow(() -> new ResponseStatusException(HttpStatus.BAD_REQUEST));
-
+        Client currentClient = this.clientRepository.findByUsername(username);
+        if (currentClient == null) return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("User not found");
         Utils.copyNonNullProperties(client, currentClient);
         Client updateClient = this.clientRepository.save(currentClient);
 
-        return ResponseEntity.ok("Client " + updateClient.getName() + " updated");
+        return ResponseEntity.ok("Client " + updateClient.getUsername() + " updated");
     }
 
-    @DeleteMapping(path = "/{id}")
-    public ResponseEntity<String> deleteClients(@PathVariable Long id) {
-        Client client = clientRepository.findById(id).orElseThrow(() -> new ResponseStatusException(HttpStatus.BAD_REQUEST));
+    @DeleteMapping(path = "/{username}")
+    public ResponseEntity<String> deleteClientsById(@PathVariable String username) {
+        Client client = clientRepository.findByUsername(username);
+        if (client == null) return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("User not found");
+
         clientRepository.delete(client);
-        return ResponseEntity.ok("Client " + client.getName() + " was deleted");
+        return ResponseEntity.ok("Client " + client.getUsername() + " was deleted");
     }
 }
